# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project-R (DeepQuant) API",
    "version": "2.0.0",
    "description": "AI-driven algorithmic trading platform with R-Factor market intelligence, Historify data management, quantitative backtesting, and AI-powered trade decisions. Built on Dhan V2 for Indian equity & F&O markets.",
    "contact": {
      "name": "Project-R"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/historify/day-chart": {
      "get": {
        "tags": ["Historify"],
        "summary": "Get intraday chart data",
        "description": "Returns intraday 5-min candles for a specific date with VWAP, EMA20, EMA50, and RSI.",
        "operationId": "getDayChart",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "exchange",
            "in": "query",
            "schema": { "type": "string", "default": "NSE" }
          },
          {
            "name": "interval",
            "in": "query",
            "schema": { "type": "string", "default": "5min" }
          }
        ],
        "responses": {
          "200": {
            "description": "Intraday chart data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "candles": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/OHLCVCandle" }
                    },
                    "indicators": {
                      "type": "object",
                      "properties": {
                        "vwap": { "type": "array", "items": { "type": "number", "nullable": true } },
                        "ema20": { "type": "array", "items": { "type": "number", "nullable": true } },
                        "ema50": { "type": "array", "items": { "type": "number", "nullable": true } },
                        "rsi": { "type": "array", "items": { "type": "number", "nullable": true } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/backtest": {
      "get": {
        "tags": ["Backtest"],
        "summary": "Get available stocks and regimes",
        "operationId": "getBacktestInfo",
        "responses": {
          "200": {
            "description": "Available stocks and regimes for backtesting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "availableStocks": { "type": "array", "items": { "type": "string" } },
                    "regimes": { "type": "array", "items": { "type": "string" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Backtest"],
        "summary": "Run mock backtest",
        "description": "Run a simplified backtest simulation with configurable R-Factor parameters and weights.",
        "operationId": "runMockBacktest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": { "type": "string", "default": "RELIANCE" },
                  "days": { "type": "integer", "default": 90 },
                  "rThreshold": { "type": "number", "default": 1.5 },
                  "volWeight": { "type": "number" },
                  "turnWeight": { "type": "number" },
                  "oiWeight": { "type": "number" },
                  "spreadWeight": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Backtest results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MockBacktestResult" }
              }
            }
          }
        }
      }
    },
...
    "/api/ai-trading/config": {
      "get": {
        "tags": ["AI Trading"],
        "summary": "Get risk configuration",
        "description": "Returns current risk config and defaults.",
        "operationId": "getAiConfig",
        "responses": {
          "200": {
            "description": "Risk config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "config": { "$ref": "#/components/schemas/RiskConfig" },
                    "defaults": { "$ref": "#/components/schemas/RiskConfig" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["AI Trading"],
        "summary": "Update risk configuration",
        "description": "Partial update of risk config. Validated: maxCapitalPerTrade 0.5-10%, maxOpenPositions 1-20, minADXThreshold 10-50.",
        "operationId": "updateAiConfig",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RiskConfig" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "config": { "$ref": "#/components/schemas/RiskConfig" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RiskConfig": {
        "type": "object",
        "description": "Risk management configuration for AI trading.",
        "properties": {
          "maxCapitalPerTrade": { "type": "number", "description": "% of total capital (default: 2%)", "default": 2 },
          "maxOpenPositions": { "type": "integer", "description": "Max simultaneous positions", "default": 5 },
          "maxDailyLoss": { "type": "number", "description": "% of capital to stop trading", "default": 5 },
          "maxSectorExposure": { "type": "number", "description": "% of capital per sector", "default": 20 },
          "defaultStopLossPct": { "type": "number", "description": "% below entry", "default": 1.5 },
          "defaultTargetPct": { "type": "number", "description": "% above entry", "default": 3 },
          "minRiskReward": { "type": "number", "description": "Minimum R:R ratio", "default": 1.5 },
          "minRFactorThreshold": { "type": "number", "description": "Min R-Factor to consider", "default": 2.0 },
          "minADXThreshold": { "type": "number", "description": "Min ADX for strong trend", "default": 28 },
          "paperTrading": { "type": "boolean", "description": "Log only, no real orders", "default": true },
          "entryWindowStart": { "type": "string", "description": "HH:MM IST", "default": "09:45" },
          "entryWindowEnd": { "type": "string", "description": "HH:MM IST", "default": "11:00" },
          "exitMode": {
            "type": "string",
            "enum": ["ai", "fixed-profit", "fixed-time", "trailing-sl"],
            "default": "fixed-profit"
          },
          "fixedProfitTarget": {
            "type": "number",
            "description": "INR profit per trade to auto-exit",
            "default": 5000
          },
          "fixedExitTime": { "type": "string", "description": "HH:MM IST force-exit time", "default": "15:10" },
          "trailingSlPct": { "type": "number", "description": "Trailing stop-loss %", "default": 1.0 }
        }
      },
...
      "OHLCVCandle": {
        "type": "object",
        "description": "OHLCV candle data point.",
        "properties": {
          "timestamp": { "type": "number", "description": "Unix timestamp in seconds" },
          "open": { "type": "number" },
          "high": { "type": "number" },
          "low": { "type": "number" },
          "close": { "type": "number" },
          "volume": { "type": "number" }
        }
      },
...
      "BacktestTrade": {
        "type": "object",
        "properties": {
          "entryDate": { "type": "string" },
          "entryPrice": { "type": "number" },
          "exitDate": { "type": "string" },
          "exitPrice": { "type": "number" },
          "shares": { "type": "integer" },
          "pnl": { "type": "number" },
          "pnlPct": { "type": "number" },
          "fees": { "type": "number" }
        }
      },
      "MockBacktestResult": {
        "type": "object",
        "description": "Simplified mock backtest result.",
        "properties": {
          "symbol": { "type": "string" },
          "regime": { "type": "string" },
          "parameters": { "type": "object" },
          "stats": {
            "type": "object",
            "properties": {
              "totalTrades": { "type": "integer" },
              "wins": { "type": "integer" },
              "losses": { "type": "integer" },
              "winRate": { "type": "number" },
              "totalReturn": { "type": "number" },
              "profitFactor": { "type": "number" },
              "maxDrawdown": { "type": "number" },
              "sharpeRatio": { "type": "number" },
              "avgWin": { "type": "number" },
              "avgLoss": { "type": "number" },
              "expectancy": { "type": "number" }
            }
          },
          "trades": { "type": "array", "items": { "$ref": "#/components/schemas/BacktestTrade" } },
          "availableStocks": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}